tyagachi/requirements.txt
pandas>=2.0.0
pyyaml>=6.0.0
requests>=2.28.0
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
